Move Id Gen tests to new folder structure
