+++ modules/kernel-tests/src/test/scala/io/janstenpickle/trace4cats/GenAssertions.scala
@@ -1,11 +1,11 @@
-package io.janstenpickle.trace4cats.model
+package io.janstenpickle.trace4cats
 
+import cats.Eq
 import cats.effect.IO
-import cats.kernel.Eq
 import cats.syntax.parallel._
 import org.scalatest.{Assertion, Assertions}
 
-private[model] object GenAssertions extends Assertions {
+private[trace4cats] object GenAssertions extends Assertions {
 
   def assertAllDistinct[A: Eq](generate: IO[A]): IO[Assertion] =
     generate
